requirements.txt
pathvalidate
slugify
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
fuzzywuzzy==0.18.0
idna==2.10
infi==0.0.1
lxml==4.5.1
Pillow==7.2.0
pycparser==2.20
PySimpleGUI==4.22.0
python-Levenshtein-wheels==0.13.1
python-magic==0.4.18
python-magic-bin==0.4.14
pytz==2020.1
requests==2.24.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.9
xmltodict==0.12.0
